requirements.txt
asgiref
boto3
botocore
click
Django
django-cors-headers
django-storages
djangorestframework
djangorestframework-simplejwt
jmespath
mccabe
mypy-extensions
packaging
pathspec
Pillow
platformdirs
pycodestyle
pyflakes
PyJWT
python-dateutil
pytz
s3transfer
six
sqlparse
tomli
typing_extensions
urllib3
gunicorn
whitenoise
drf-yasg
